backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1